moved choose_filter_helper from main.py to filters.py
